refactor use cases and ui components to improve modularity and naming consistency
