+++ src/main/kotlin/Main.kt
@@ -1,6 +1,7 @@
 package org.example
 
 import org.example.di.appModule
+import org.example.di.uiModule
 import org.example.di.useCasesModule
 import org.example.presentaion.GetFoodUi
 import org.koin.core.context.startKoin
@@ -8,7 +9,7 @@ import org.koin.java.KoinJavaComponent.getKoin
 
 fun main() {
     startKoin {
-        modules(appModule,useCasesModule)
+        modules(appModule,useCasesModule,uiModule)
     }
 
     val getFood = getKoin().get<GetFoodUi>()
