+++ src/main/kotlin/di/AppModule.kt
@@ -4,8 +4,6 @@ import org.example.data.CsvFileReader
 import org.example.data.CsvFoodRepository
 import org.example.data.FoodCsvFileParse
 import org.example.domain.repository.FoodRepository
-import org.example.presentaion.GetEasyFoodSuggestionUi
-import org.example.presentaion.GetFoodUi
 import org.koin.dsl.module
 import java.io.File
 
@@ -14,7 +12,4 @@ val appModule = module {
     single { CsvFileReader(get()) }
     single { FoodCsvFileParse() }
     single<FoodRepository> { CsvFoodRepository(get(), get()) }
-
-    single { GetFoodUi(get()) }
-    single { GetEasyFoodSuggestionUi(get()) }
 }
