+++ src/main/kotlin/di/UiModule.kt
@@ -1,5 +1,6 @@
 package org.example.di
 
+import org.example.presentaion.GetEasyFoodSuggestionUi
 import org.example.presentaion.GetFoodUi
 import org.example.presentaion.GuessGameUi
 import org.koin.dsl.module
@@ -7,4 +8,6 @@ import org.koin.dsl.module
 val uiModule = module {
     single { GetFoodUi(get()) }
     single { GuessGameUi(get()) }
+    single { GetFoodUi(get()) }
+    single { GetEasyFoodSuggestionUi(get()) }
 }
