+++ src/main/kotlin/di/UseCasesModule.kt
@@ -2,12 +2,11 @@ package org.example.di
 
 import org.example.domain.usecase.GetEasyFoodSuggestionUseCase
 import org.example.domain.usecase.GetFoodUseCase
-import org.example.domain.usecase.GetFoodUC
-import org.example.domain.usecase.GuessGameUC
+import org.example.domain.usecase.GuessGameUseCase
 import org.koin.dsl.module
 
 val useCasesModule = module {
     single { GetFoodUseCase(get()) }
     single { GetEasyFoodSuggestionUseCase(get()) }
-    single { GuessGameUC(get()) }
+    single { GuessGameUseCase(get()) }
 }
